edit

+++ db/migrate/10200113094418_devise_create_users.rb
@@ -15,7 +15,7 @@ class DeviseCreateUsers < ActiveRecord::Migration[5.2]
       t.integer :birthyear, null: false
       t.integer :birthmonth, null: false
       t.integer :birthday, null: false
-      t.integer :phonenumber, null: false
+      t.bigint :phonenumber, null: false
 
       ## Recoverable
       t.string   :reset_password_token
